memory_display: Fix type of argument passed to %*s
On AArch64, pointer arithmetic in (pos - line) resolves into "long
int", whereas "%*s" is expecting regular "int". Add explicit cast to
avoid compiler warnings.

+++ common/memory_display.c
@@ -96,7 +96,7 @@ int __pr_memory_display(int level, const void *addr, loff_t offs, unsigned nbyte
 			offs += size;
 		}
 
-		pos += sprintf(pos, "%*s", 61 - (pos - line), "");
+		pos += sprintf(pos, "%*s", (int)(61 - (pos - line)), "");
 
 		cp = linebuf;
 		for (i = 0; i < linebytes; i++) {
